alx_travel_app/requirement.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.1
celery>=5.3.6
drf-yasg>=1.21.7
